requirements.txt
absl-py==0.13.0
matplotlib==3.3.4
ml-collections==0.1.0
numpy==1.20.2
torch==1.9.0
pytorch-lightning==1.3.8
scikit-learn==0.24.2
scipy==1.7.0
torchdiffeq==0.2.2
h5py==3.2.1
sortedcontainers==2.1.0
ipython==7.22.0
